fix (df_read_csv.c) fixed some reading bugs

+++ src/basics_operations/df_dump.c
@@ -5,20 +5,21 @@
 ** df_dump
 */
 
+#include <stdio.h>
 #include "dataframe.h"
 
 void df_dump(dataframe_t *dataframe)
 {
-    for (int i = 0; i < dataframe->nb_columns; i++) {
+    for (size_t i = 0; i < dataframe->nb_columns; i++) {
         printf("| %-10s ", dataframe->columns[i]->name);
     }
     printf("|\n");
-    for (int i = 0; i < dataframe->nb_columns; i++) {
+    for (size_t i = 0; i < dataframe->nb_columns; i++) {
         printf("|------------");
     }
     printf("|\n");
-    for (int e = 0; e < dataframe->nb_rows; e++) {
-        for (int i = 0; i < dataframe->nb_columns; i++) {
+    for (size_t e = 0; e < dataframe->nb_rows; e++) {
+        for (size_t i = 0; i < dataframe->nb_columns; i++) {
             printf("| %-10s ", dataframe->columns[i]->column_content[e]);
         }
         printf("|\n");
